requirements.txt
gymnasium[atari]
ale-py
autorom
numpy
torch
pyyaml
tensorboard
pandas
tqdm
pytest
